Rename Footer to Pagination
