+++ packages/components/src/DataTable/Pagination.css.d.ts
